Adding player selection
I have started to add choices for the player on whether or not to draw more or not. The 2nd option(not drawing) is not done yet. The card deck was changed from a number set to a list. I don't work on lists often so I had to look at some examples of card games in Python. I will try to make it look more readable and not have so many nested if statements as projects like these are prone to have.